chore: move primitve checks to the typed util
