+++ __tests__/utils/num.test.ts
@@ -1,7 +1,6 @@
 import {
   isHex,
   toBigInt,
-  isBigInt,
   toHex,
   hexToDecimalString,
   cleanHex,
@@ -15,8 +14,6 @@ import {
   toCairoBool,
   hexToBytes,
   addPercent,
-  isNumber,
-  isBoolean,
 } from '../../src/utils/num';
 import { num } from '../../src';
 
@@ -46,23 +43,6 @@ describe('toBigInt', () => {
 
   test('should throw for invalid arg', () => {
     expect(() => toBigInt('test')).toThrow();
-  });
-});
-
-describe('isBigInt', () => {
-  test('should return true for big integers', () => {
-    expect(isBigInt(BigInt(10))).toBe(true);
-    expect(isBigInt(BigInt('9007199254740991'))).toBe(true);
-  });
-
-  test('should return false for non-big integers', () => {
-    expect(isBigInt(10)).toBe(false);
-    expect(isBigInt('10')).toBe(false);
-    expect(isBigInt(undefined)).toBe(false);
-    expect(isBigInt(null)).toBe(false);
-    expect(isBigInt({})).toBe(false);
-    expect(isBigInt([])).toBe(false);
-    expect(isBigInt(true)).toBe(false);
   });
 });
 
@@ -174,39 +154,6 @@ describe('addPercent', () => {
     expect(addPercent(200, -50)).toBe(100n);
     expect(addPercent(200, -100)).toBe(0n);
     expect(addPercent(200, -150)).toBe(-100n);
-  });
-});
-
-describe('isNumber', () => {
-  test('should correctly determine if value is a number', () => {
-    expect(isNumber(0)).toBe(true);
-    expect(isNumber(123)).toBe(true);
-    expect(isNumber(-123)).toBe(true);
-
-    expect(isNumber(123n)).toBe(false);
-    expect(isNumber('')).toBe(false);
-    expect(isNumber('123')).toBe(false);
-    expect(isNumber(true)).toBe(false);
-    expect(isNumber(false)).toBe(false);
-    expect(isNumber(null)).toBe(false);
-    expect(isBoolean([])).toBe(false);
-    expect(isBoolean({})).toBe(false);
-  });
-});
-
-describe('isBoolean', () => {
-  test('should correctly determine if value is a boolean', () => {
-    expect(isBoolean(true)).toBe(true);
-    expect(isBoolean(false)).toBe(true);
-
-    expect(isBoolean(0)).toBe(false);
-    expect(isBoolean(1)).toBe(false);
-    expect(isBoolean('')).toBe(false);
-    expect(isBoolean('true')).toBe(false);
-    expect(isBoolean('false')).toBe(false);
-    expect(isBoolean(null)).toBe(false);
-    expect(isBoolean([])).toBe(false);
-    expect(isBoolean({})).toBe(false);
   });
 });
