+++ src/utils/num.ts
@@ -4,6 +4,7 @@ import { BigNumberish } from '../types';
 import assert from './assert';
 import { addHexPrefix, buf2hex, removeHexPrefix } from './encode';
 import { MASK_31 } from '../constants';
+import { isNumber, isBigInt } from './typed';
 
 /** @deprecated prefer importing from 'types' over 'num' */
 export type { BigNumberish };
@@ -42,24 +43,6 @@ export function isHex(hex: string): boolean {
  */
 export function toBigInt(value: BigNumberish): bigint {
   return BigInt(value);
-}
-
-/**
- * Test if value is bigint
- *
- * @param value value to test
- * @returns {boolean} true if value is bigint, false otherwise
- * @example
- * ```typescript
- * isBigInt(10n); // true
- * isBigInt(BigInt('10')); // true
- * isBigInt(10); // false
- * isBigInt('10'); // false
- * isBigInt(null); // false
- * ```
- */
-export function isBigInt(value: any): value is bigint {
-  return typeof value === 'bigint';
 }
 
 /**
@@ -338,44 +321,6 @@ export function hexToBytes(str: string): Uint8Array {
 export function addPercent(number: BigNumberish, percent: number): bigint {
   const bigIntNum = BigInt(number);
   return bigIntNum + (bigIntNum * BigInt(percent)) / 100n;
-}
-
-/**
- * Check if a value is a number.
- *
- * @param {unknown} value - The value to check.
- * @returns {boolean} Returns true if the value is a number, otherwise returns false.
- * @example
- * ```typescript
- * const result = isNumber(123);
- * // result = true
- *
- * const result2 = isNumber("123");
- * // result2 = false
- * ```
- * @return {boolean} Returns true if the value is a number, otherwise returns false.
- */
-export function isNumber(value: unknown): value is number {
-  return typeof value === 'number';
-}
-
-/**
- * Checks if a given value is of boolean type.
- *
- * @param {unknown} value - The value to check.
- * @returns {boolean} - True if the value is of boolean type, false otherwise.
- * @example
- * ```typescript
- * const result = isBoolean(true);
- * // result = true
- *
- * const result2 = isBoolean(false);
- * // result2 = false
- * ```
- * @return {boolean} - True if the value is of boolean type, false otherwise.
- */
-export function isBoolean(value: unknown): value is boolean {
-  return typeof value === 'boolean';
 }
 
 /**
